+++ pkg/nginxpm/letsencrypt_certificate_test.go
@@ -150,7 +150,7 @@ func TestFindLetEncryptCertificateByID(t *testing.T) {
 
 	tests := []struct {
 		name           string
-		certificateID  uint16
+		certificateID  int
 		serverResponse []LetsEncryptCertificate
 		expectedCert   *LetsEncryptCertificate
 		serverStatus   int
